Changed get_used_widget function to another file
